修复安装程序 bug
修复安装程序的一些问题

+++ index.php
@@ -5,7 +5,7 @@ $pathinfo = pathinfo($_SERVER['PHP_SELF']);
 $path = str_replace("/" . $pathinfo['basename'], "", $_SERVER['PHP_SELF']);
 define("ROOT", str_replace("\\", "/", __DIR__));
 define("DOCROOT", $path);
-if(!file_exists(ROOT . "/include/data/config.php")) {
+if(!file_exists(ROOT . "/include/data/config.php") && !isset($_GET['installed'])) {
 	echo "<script>location='install/';</script>";
 	exit;
 }
